Gates

+++ app/Http/Controllers/Admin/UserController.php
@@ -21,6 +21,8 @@ class UserController extends BaseController
     {
         parent::__construct();
 
+        $this->middleware('auth');
+
         $this->userRepository = app(UserRepository::class);
         $this->cityRepository = app(CityRepository::class);
     }
@@ -53,7 +55,11 @@ class UserController extends BaseController
     public function store(UserCreateRequest $request)
     {
         $data = $request->input();
-        $item = (new User())->create($data);
+        if(\Gate::allows('update-user')) {
+            $item = (new User())->create($data);
+        } else {
+            $item = null;
+        }
 
         if ($item) {
             return redirect()->route('admin.user.edit', [$item->id])
@@ -108,7 +114,11 @@ class UserController extends BaseController
         }
 
         $data = $request->input();
-        $item->update($data);
+        if(\Gate::allows('update-user', $id)) {
+            $item->update($data);
+        } else {
+            $item = null;
+        }
 
         if ($item) {
             return redirect()->route('admin.user.edit', [$item->id])
@@ -130,8 +140,11 @@ class UserController extends BaseController
         if($id == 1 )
             return back()
                 ->withErrors(['msg' => 'Ошибка удаления']);
-
-        $result = User::destroy($id);
+        if(\Gate::allows('delete-user', $id)) {
+            $result = User::destroy($id);
+        } else {
+            $result = null;
+        }
 
         if($result) {
             return redirect()
